pyproject.toml
[tool.poetry]
name = "audio_number_classification"
version = "0.1.0"
description = ""
authors = ["Olga Golovkina"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.1.2"
pandas = "^2.1.4"
torchaudio = "^2.1.2"
hydra-core = "^1.3.2"
tqdm = "^4.66.1"
gitpython = "^3.1.40"
dvc = "^3.38.1"
dvc-gdrive = "^3.0.1"
soundfile = "^0.12.1"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
